[ONECTA] Creation of new binding connecting to Daikin Cloud Onecta.
With the newer Daikin units it is no longer possible to control them directly. The units can only be connected to the Daikin cloud called Onecta.
The units can then ‘only’ be controlled with the Onecta app on a phone or tablet.
This binding makes it possible to still control the units with OpenHAB. It’s now done by connecting the binding to Daikin’s Onecta.
The unit information is then received from the Daikin cloud just like the app. Commands to the units also run via the Daikin Cloud.
Older units can also be controlled with this binding as long as they are registered in Onecta.

This Binding is already in use by several users.
https://community.openhab.org/t/daikin-onecta-cloud-binding-4-0-0-0-5-0-0-0

Signed-off-by: Alexander Drent <Alex@Drent-ict.nl>